delete authorization

+++ App/controllers/ListingController.php
@@ -5,6 +5,8 @@ namespace App\Controllers;
 use ErrorException;
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -26,7 +28,7 @@ class ListingController
         //inspectAndDie(Validation::string('r23', 3, 5));
         //inspectAndDie(Validation::email('123@test.com'));
         //inspectAndDie(Validation::match('a', 'a'));
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', [
             'listings' => $listings
@@ -75,7 +77,7 @@ class ListingController
         // Turn value into key. --array_flip()
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -142,9 +144,16 @@ class ListingController
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if listing exists
         if (!$listing) {
             ErrorController::notfound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+            return redirect('/listings/' . $listing->$id);
         }
 
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
